Try to use non-blocking read()'s.
This assumes that display() is much faster than audioRead(). If this
holds true, then we can render some additional frames while the audio
buffer is not yet filled. In turn, this results in a "smoother" user
experience -- especially when using the overtone guidelines.

As a bonus, this allows you to use a larger audio buffer without getting
a "slide show". Larger audio buffers mean larger latency but also a
higher accuracy of the spectrogram.