Fixed wrong array size for fftw.out.
According to the docs, the output array must be of size "in / 2 + 1".

The old code only alloc'd "in / 2" values which could lead to freezes or
crashes (in certain situations).

"fftw.outlen", however, must remain the same due to the required texture
size of 2^n. We simply don't read that new extra value.