Tolerance for changing the label's position.
Just think of a user who positions the mouse pointer at exactly 50% of
the screen. One pixel to the left and the label is flipped -- two pixels
to the right and it's flipped again! That's annoying.

This commit adds a "dead zone" of some percent around the screen center.
Inside that zone, no label flipping is done.