Code style (line continuation, alignment).
Tabs set the indentation level of a block. If a long is too long and
must be continued on the next line, then use the same amount of tabs at
the beginning of the line. After them, use as many spaces as needed to
reach the last "parenthesis level".

tl;dr: Tabs for indentation, spaces for aligning.